feat: Update S3FileSystem to support readable proxy URLs and allow acl to be null
This commit introduces updates to the S3FileSystem to support readable proxy URLs and allows the acl parameter to be set to null. This change addresses issues where certain storage systems would throw errors if an unsupported header value was provided. Additionally, a new `$readableProxyUrl` property has been added to improve the management and readability of proxy URLs.